+++ src/main/java/com/jcraft/jsch/IdentityFile.java
@@ -45,7 +45,7 @@ class IdentityFile implements Identity {
     return new IdentityFile(name, kpair);
   }
 
-  private IdentityFile(String name, KeyPair kpair) throws JSchException {
+  private IdentityFile(String name, KeyPair kpair) {
     this.identity = name;
     this.kpair = kpair;
   }
@@ -95,29 +95,20 @@ class IdentityFile implements Identity {
   }
 
   /**
-   * @deprecated This method should not be invoked.
-   * @see #setPassphrase(byte[] passphrase)
-   */
-  @Override
-  @Deprecated
-  public boolean decrypt() {
-    throw new RuntimeException("not implemented");
-  }
-
-  /**
    * Returns the name of the key algorithm.
    *
-   * @return "ssh-rsa" or "ssh-dss"
+   * @return the name of the key algorithm
    */
   @Override
   public String getAlgName() {
-    byte[] name = kpair.getKeyTypeName();
-    return Util.byte2str(name);
+    return kpair.getKeyTypeString();
   }
 
   /**
    * Returns the name of this identity. It will be useful to identify this object in the
    * {@link IdentityRepository}.
+   *
+   * @return the name of this identity
    */
   @Override
   public String getName() {